+++ packages/nextjs/app/categories/page.tsx
@@ -1,5 +1,5 @@
-import { Breadcrumbs } from "components/Breadcrumbs";
-import { CategoryList } from "components/CategoryList";
+import { Breadcrumbs } from "app/components/Breadcrumbs";
+import { CategoryList } from "app/components/CategoryList";
 import { Metadata } from "next";
 import { WeDontSellBreadBanner } from "../ui/shared-ui";
 import { getAllCategories } from "utils/getAllCategoriesQuery";
