+++ packages/nextjs/app/components/ImageCarousel.tsx
@@ -1,6 +1,6 @@
 import { ProductImage } from "utils/groqTypes/ProductList";
 import * as React from "react";
-import { Image } from "components/Image";
+import { Image } from "app/components/Image";
 import { ImageCarousel as BaseImageCarousel } from "../ui/shared-ui";
 
 export type ImageCarouselProps = {
