+++ packages/nextjs/app/components/ProductFilters/FilterGroup.tsx
@@ -3,7 +3,7 @@
 import type { FilterGroup as FilterGroupType } from "utils/filters";
 import * as React from "react";
 import { ChangeEvent } from "react";
-import { Checkbox } from "../../app/ui/shared-ui";
+import { Checkbox } from "../../ui/shared-ui";
 import { useRouterQueryParams } from "utils/useRouterQueryParams";
 
 type FilterGroupProps = {
