+++ packages/nextjs/app/layout.tsx
@@ -1,8 +1,8 @@
 import { Footer, MobileNavProvider } from "./ui/shared-ui";
 import "./global.css";
-import { Header } from "components/Header/Header";
+import { Header } from "app/components/Header/Header";
 import { Metadata } from "next";
-import { CartProvider } from "components/CartContext";
+import { CartProvider } from "app/components/CartContext";
 import { AnimatePresence, MotionConfig } from "./ui/framer";
 import localFont from "next/font/local";
